Move LocalFileGzip to collab server package
